(fixes) test config issues
