+++ jest.config.js
@@ -1,5 +1,9 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  coverageProvider: 'babel'
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	coverageProvider: 'babel',
+	testMatch: ['**/__tests__/**/*.+(ts)', '**/?(*.)+(spec|test).+(ts)'],
+	transform: {
+		'^.+\\.(ts)$': 'ts-jest',
+	},
 };
