--- src/test/sum.ts
@@ -1,3 +1,0 @@
-export default function sum(a: number, b: number) {
-  return a + b;
-}
